pare down exposed methods on round
